refactor code for readability and easier extensibility of endpoints and handlers
