+++ restapi/endpoints/endpoints.go
@@ -18,23 +18,23 @@ const (
 	timeout     = 1
 )
 
+type RestAPI struct {
+	HelloWorlder
+	HealthChecker
+}
+
+type HelloWorlder interface {
+	HelloWorld([]string) (string, error)
+}
+
 type HealthChecker interface {
 	HealthCheck() int
 }
 
-type RestAPIHealthCheckEndpoint struct{}
+// resources for servicing hello world endpoint
+type HelloWorldEndpoint struct{}
 
-func (api RestAPIHealthCheckEndpoint) HealthCheck() int {
-	return http.StatusOK
-}
-
-type HealthWorlder interface {
-	HelloWorld([]string) (string, error)
-}
-
-type RestAPIHelloWorldEndpoint struct{}
-
-func (api RestAPIHelloWorldEndpoint) HelloWorld(names []string) (string, error) {
+func (api HelloWorldEndpoint) HelloWorld(names []string) (string, error) {
 	// Set up a connection to the server.
 	conn, err := grpc.Dial(addressDB, grpc.WithInsecure(), grpc.WithBlock(), grpc.WithTimeout(timeout*time.Second))
 	if err != nil {
@@ -64,4 +64,32 @@ func (api RestAPIHelloWorldEndpoint) HelloWorld(names []string) (string, error) 
 	}
 
 	return msg, grpcErr
+}
+
+func (api *RestAPI) HelloWorldHandler(w http.ResponseWriter, r *http.Request) {
+	log.Println("r.URL.Query()[\"name\"]", r.URL.Query()["name"])
+
+	msg, err := api.HelloWorld(r.URL.Query()["name"])
+	if err != nil {
+		log.Println(err)
+		//TODO: write user friendly error message
+		http.Error(w, http.StatusText(http.StatusInternalServerError)+": "+err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	log.Printf("%s?%s; Message:%s", r.URL.Path, r.URL.RawQuery, msg)
+	fmt.Fprint(w, msg)
+}
+
+// resources for servicing health check endpoint
+type HealthCheckEndpoint struct{}
+
+func (api HealthCheckEndpoint) HealthCheck() int {
+	return http.StatusOK
+}
+
+func (api *RestAPI) HealthCheckHandler(w http.ResponseWriter, r *http.Request) {
+	w.WriteHeader(api.HealthCheck())
+	fmt.Fprint(w, "OK")
+	log.Println(r.URL.Path, http.StatusOK)
 }
